+++ web/src/App.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
 
-import RootWrapper from './components/rootWrapper/rootWrapper'
-import Frontpage from "./components/selection/frontpage";
 import BrowserBroadcaster from "./components/broadcast/Broadcast";
 import PlayerPage from "./components/player/PlayerPage";
+import RootWrapper from "./components/rootWrapper/RootWrapper";
+import Frontpage from "./components/selection/Frontpage";
 
 function App() {
   return (
